attributes tests refactoring
