+++ spec/exec.js
@@ -26,7 +26,7 @@ require("./special-cases");
 require("./template");
 require("./escape-html");
 require("./styles");
-require("./attrs");
+require("./render-attrs");
 require("./adler32");
 require("./render-error");
 require("./elements");
