+++ spec/render-attrs.js
